Eclipse errors cleanup
-Added ignore to non-supportable antrun task in maven lifecycle
-fixed what should be static references
-removed exception declarations from methods that don't throw that exception

+++ src/main/java/org/oscarehr/PMmodule/model/ProgramClientRestriction.java
@@ -147,7 +147,7 @@ public class ProgramClientRestriction implements Serializable {
     }
 
     public int hashCode() {   
-        return (int) (id ^ (id >>> 32));
+        return (id ^ (id >>> 32));
     }
 
 	public String getCommentId() {
